% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{create_table1}
\alias{create_table1}
\title{Create descriptive tables}
\usage{
create_table1(
  df,
  headvar,
  rowvars,
  headvar_levels = NULL,
  levels_order = NULL,
  rowvar_labels = NULL,
  headvar_na_level = "Missing",
  file_name = "mytable",
  header = "add span row title",
  digits_num = 1L,
  do_test = FALSE,
  col_total = FALSE,
  numvar_stats = c("meansd", "Nmiss"),
  catvar_stats = c("countrowpct"),
  stats_labels = list(Nmiss = "(Missing)"),
  out_type = "docx",
  font_size = 11,
  font_family = "Arial",
  ...
)
}
\arguments{
\item{df}{a dataframe}

\item{headvar}{The variable name that you want your descriptive statistics stratified by (outcome or exposure variable).}

\item{rowvars}{a vector explanatory variables (in the order you want them to appear in the table)
example: c("age", "sex", "income")}

\item{headvar_levels}{labels to be used for each level of head variable
(eg: c("Yes"= "1","No"= "0", "Missing"= NA))}

\item{levels_order}{order of columns from left to right (eg: c("No", "Yes"))}

\item{rowvar_labels}{Labels for explanatory variables (in correct order)
example: c("Age (yr)", "Gender", "Monthly income")}

\item{headvar_na_level}{How to name NA level of head variable (Eg: "Missing")}

\item{file_name}{Name of the .docx file contain the table}

\item{header}{Spanning header for the head variable columns}

\item{digits_num}{Number of digits for numerical values}

\item{do_test}{TRUE if you want to perform statistical analyses (eg: "chi", "anova", "kwt", ..etc)
can specify tets with additional arguments (?arsenal::tableby.control)}

\item{col_total}{Total for columns if TRUE}

\item{numvar_stats}{Stats to display for numerical variables}

\item{catvar_stats}{Stats to display for categorical variables}

\item{stats_labels}{List of labels for respective stats}

\item{out_type}{select "pdf" or "docx"}

\item{font_size}{change font size (default 11)}

\item{font_family}{change font family "Arial" is default}

\item{...}{Takes any argument for the arsenal::tableby.control() function}
}
\value{
A .docx document with the table
}
\description{
Create descriptive tables
}
\examples{
 df <- mtcars

create_table1(df,headvar= "am",
                 headvar_levels= c("Yes"= "1","No"= "0") ,
                 rowvars= c("wt","gear","carb"),
                 levels_order= c("Yes","No"),
                 rowvar_labels= c("Weight","Gear","Number of carbs"))
}
