% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{tableby_to_flextable}
\alias{tableby_to_flextable}
\title{Convert tableby object to a flextable}
\usage{
tableby_to_flextable(
  tblby,
  header_row = "add span row title",
  colwidths = c(1, 4),
  size = 11,
  family = "Arial"
)
}
\arguments{
\item{tblby}{A tableby object}

\item{header_row}{title for header row}

\item{colwidths}{based on the headvar}

\item{size}{font size}

\item{family}{font family}
}
\value{
none
}
\description{
Convert tableby object to a flextable
}
